fix(golden): always show escaped golden files when comparing the outputs
This change makes `RequireEqual` always escape the control codes and
escape sequences from the diff output when comparing the golden file
with the output.

It also deprecates the `RequireEqualEscape` function, which was
previously used to escape the control codes and escape sequences from
the output before comparing it with the golden file.