feat(example): implement a faketty example
This implements a faketty executable that can be used to run a command
in a fake terminal. This is useful to force escape sequences to be
printed to the terminal while piping the output to another command like
`sequin`.

It also takes the `rows` and `cols` flags to set the terminal size and
uses 80x24 as the default.