feat(ansi): XTWINOPS and WindowOp functions
This implements the XTWINOPS and WindowOp functions that allow
manipulating the terminal window. Using `ansi.WindowOp` with
`ReportWindowSizeWinOp`, we can query the terminal for the pixel size of
the window. This is useful for rendering images in the terminal and
calculate the cell size based on how many pixels and cells are available.

See https://invisible-island.net/xterm/ctlseqs/ctlseqs.html#h4-Functions-using-CSI-_-ordered-by-the-final-character-lparen-s-rparen:CSI-Ps;Ps;Ps-t.1EB0