feat(input): support kitty graphics responses
This adds support for parsing Kitty Graphics Protocol responses. This is
useful for applications that want to use the Kitty terminal's graphics
protocol to render images and other graphics.

When using the protocol, the terminal responds on querying and/or
transmitting images. This parses these responses and emits events for
them.