feat(ansi): define Method and make wcwidth aware variants (#341)
* feat(ansi): define Method and make wcwidth aware variants

This commit introduces a new type, Method, that represents how the
renderer should calculate the display width of cells. It adds
helper methods to the Method type that allow users to use either
WcWidth, which uses go-runewidth, or GraphemeWidth, which uses
uniseg, to calculate the width of strings and text.

Methods like `ansi.StringWidth`, `ansi.Truncate`, `ansi.TruncateLeft`,
and `ansi.Cut` now have variants that are aware of wide characters
and runes. These new methods are prefixed with `Wc` and use
go-runewidth to calculate the width of strings.

Related: https://github.com/charmbracelet/x/pull/217

* feat(ansi): add WcStringWidth function test

* refactor: rename wc functions to be more consistent