fix(cellbuf): break down xterm-like optimizations into capabilities
This commit breaks down the xterm-like optimizations into a set of
capabilities that can be used to determine whether a terminal
supports a specific ANSI escape sequence. This allows for more
fine-grained control over the optimizations and makes it easier to
add support for new terminals in the future.

Right now, we only add support for capabilities that are known
to be different among xterm like terminals.