perf(cellbuf): reuse ansi.Parser instances in PenWriter
The parsers used in PenWriter have very large internal buffers, but were
created from scratch for each writer.

In crush this led to a lot of allocation churn, which had a noticeable
performance impact.

The fix is to use the ansi package's pooled parsers, as they also use
large data buffers.

Because the parser is stored in the type, it must be released
afterwards. For this reason the `Close` method is extended
appropriately, and should be used everywhere.

Note that **not** calling `Close` is not a correctness issue: The
program will still run correctly, no parsers will leak, it'll just be
slower.