fix(ansi): double width multiple newlines wrapping
This fixes a bug where when we wrap mixed east asian and western text,
incorrect newlines can be inserted causing extra blank lines in the
output. However, this _slightly_ changes the wrapping behavior and cause
it to wrap a bit earlier in some cases where a double-width character
would exactly fit at the end of a line.