fix(vt): always use grapheme width using ansi.GraphemeWidth
The virtual emulator will always use ansi.GraphemeWidth to calculate the
width of grapheme clusters, regardless of whether Unicode mode is
enabled or not. This change ensures consistent and accurate width
calculations for all grapheme clusters and it's up to the caller to
decide how to handle Unicode vs non-Unicode modes by modifying the cell
buffer.